Add resize method to support dynamic resizing
This is a no-frills approach and perhaps the most flexible right now. Possible improvements would be do some of the requested size validation within a protected method in the provider class which then calls the public `resize()` method that can be overridden. Not sure how much value that would really add versus validating in the filters as this PR does currently.