Improve duplicate URL base fixing function
Uses regex to only match the last instance of `http://` or `https://` onwards, in some cases the attachment guid could be different to the individual attachment metadata sizes file name - which for AMF assets are fully qualified URLs. This means the intermediate size URLs do not need to match the original asset URL path.